requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
transformers==4.35.0
trimesh==4.0.5
rembg
huggingface-hub
imageio[ffmpeg]
gradio
xatlas==0.0.9
moderngl==5.10.0
pytorch-lightning==2.1.2
gradio
huggingface-hub
einops
omegaconf
torchmetrics
webdataset
accelerate
tensorboard
PyMCubes
trimesh
rembg
diffusers==0.20.2
bitsandbytes
imageio[ffmpeg]
xatlas
plyfile
git+https://github.com/NVlabs/nvdiffrast/
